Remove FirstStartInvite page and put it into page 2 = login
