+++ app/src/main/java/org/gdg/frisbee/android/onboarding/FirstStartStep2Fragment.java
@@ -62,6 +62,10 @@ public class FirstStartStep2Fragment extends BaseFragment {
         listener.onSkippedSignIn();
     }
 
+    void updateSender(String sender) {
+
+    }
+
     public interface Step2Listener {
         void onSignedIn();
